requirements.txt
glob2
jiwer
numpy
opencv-python
Pillow
PyMuPDF
pyocr